+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/overlay/SoAMessages.java
@@ -11,10 +11,8 @@ import net.minecraft.util.Mth;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.client.event.ClientTickEvent;
 import online.kingdomkeys.kingdomkeys.client.ClientUtils;
-import online.kingdomkeys.kingdomkeys.util.Utils;
 import online.kingdomkeys.kingdomkeys.util.Utils.Title;
 
-import javax.swing.text.Style;
 import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.Queue;